+++ typed_sql/test/typed_sql/expr/datetime_test.dart
@@ -51,6 +51,16 @@ final _cases = [
     expr: toExpr(epoch).equals(toExpr(today)),
     expected: false,
   ),
+  (
+    name: 'epoch.equals(null)',
+    expr: toExpr(epoch).equals(toExpr(null)),
+    expected: null,
+  ),
+  (
+    name: 'today.equals(null)',
+    expr: toExpr(today).equals(toExpr(null)),
+    expected: null,
+  ),
 
   // Test for .equalsValue
   (
@@ -82,6 +92,16 @@ final _cases = [
     name: 'epoch.equalsValue(today)',
     expr: toExpr(epoch).equalsValue(today),
     expected: false,
+  ),
+  (
+    name: 'epoch.equalsValue(null)',
+    expr: toExpr(epoch).equalsValue(null),
+    expected: null,
+  ),
+  (
+    name: 'today.equalsValue(null)',
+    expr: toExpr(today).equalsValue(null),
+    expected: null,
   ),
 
   // Test for .notEquals
@@ -441,7 +461,7 @@ void main() {
       final result = await db.select(
         (c.expr,),
       ).fetch();
-      check(result).isNotNull().equals(c.expected);
+      check(result).equals(c.expected);
     });
   }